requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.6
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.5.0
typing-extensions==4.12.2
djoser==2.2.3
pytelegrambotapi==4.22.1